Fix GIT#1, IOException when period is used in section name
Section name is translated to file name by nssym2path, which replaces period
to path separator. So the file creation fails for the section and throws the
exception. The solution is to replace periods to underscore like it's done
for any invalid chars. The solution is applied to not only section name but
also category and title names.